requirements.txt
streamlit
numpy
sentence-transformers
google-generativeai
PyPDF2
python-docx
spacy
plotly
pandas